validatorWorker: split out fees.js
