+++ services/validatorWorker/follower.js
@@ -2,7 +2,8 @@ const assert = require('assert')
 const isEqual = require('lodash.isequal');
 const db = require('../../db')
 const { persistAndPropagate } = require('./lib/propagation')
-const { isValidRootHash, toBNMap, getBalancesAfterFeesTree, toBNStringMap } = require('./lib')
+const { getBalancesAfterFeesTree } = require('./lib/fees')
+const { isValidRootHash, toBNMap, toBNStringMap } = require('./lib')
 const { isValidTransition, isHealthy } = require('./lib/followerRules')
 const producer = require('./producer')
 const { heartbeatIfNothingNew } = require('./heartbeat')
