+++ test/integration.js
@@ -3,7 +3,8 @@ const tape = require('tape')
 const fetch = require('node-fetch')
 const BN = require('bn.js')
 const { Channel, MerkleTree } = require('adex-protocol-eth/js')
-const { getStateRootHash, getBalancesAfterFeesTree, toBNStringMap } = require('../services/validatorWorker/lib')
+const { getStateRootHash, toBNStringMap } = require('../services/validatorWorker/lib')
+const { getBalancesAfterFeesTree } = require('../services/validatorWorker/lib/fees')
 const dummyAdapter = require('../adapters/dummy')
 
 const cfg = require('../cfg')
